Explode data along a column

Exploding an array column of length N will replicate the top level record N times. The ith replicated record will contain a struct (not an array) corresponding to the ith element of the exploded array. Exploding will not promote any fields or otherwise change the schema of the data.

sdf_explode(x, column, keep_all = FALSE)

sdf_explode_(x, column, keep_all = FALSE)

Arguments

x
An object (usually a spark_tbl) coercable to a Spark DataFrame.

column
The column to explode

keep_all
Logical. If FALSE then records where the exploded value is empty/null will be dropped.

Details

Two types of exploding are possible. The default method calls the scala explode method. This operation is supported in both Spark version > 1.6. It will however drop records where the exploding field is empty/null. Alternatively keep_all=TRUE will use the explode_outer scala method introduced in spark 2 to not drop any records.